requirements.txt
# /requirements.txt

transformers
torch
pandas
matplotlib
seaborn
tabulate
requests
python-dotenv
emoji
scikit-learn
datasets
accelerate
evaluate